qca-wifi: Enable postNOL support for 11BEA modes
While trying to switch to a configured postNOL channel, only
11AC and 11AXA modes are checked. Enable postNOL support for
11BEA modes as well.

Also enable postNOL support for 11BE 320MHz.

CRs-Fixed: 3200726

+++ umac/dfs/core/src/dfs_misc.h
@@ -40,7 +40,7 @@ void dfs_set_postnol_freq(struct wlan_dfs *dfs, qdf_freq_t postnol_freq);
  * @dfs: Pointer to wlan_dfs object.
  * @postnol_mode: PostNOL frequency value configured by the user.
  */
-void dfs_set_postnol_mode(struct wlan_dfs *dfs, uint8_t postnol_mode);
+void dfs_set_postnol_mode(struct wlan_dfs *dfs, uint16_t postnol_mode);
 
 /**
  * dfs_set_postnol_cfreq2() - DFS API to set postNOL secondary center frequency.
@@ -78,7 +78,7 @@ dfs_set_postnol_freq(struct wlan_dfs *dfs, qdf_freq_t postnol_freq)
 }
 
 static inline void
-dfs_set_postnol_mode(struct wlan_dfs *dfs, uint8_t postnol_mode)
+dfs_set_postnol_mode(struct wlan_dfs *dfs, uint16_t postnol_mode)
 {
 }
